introduce SymbolTable interface so other ST may implement it
